backend/requirements.txt
# Backend Requirements for Elasticsearch Stack

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Elasticsearch
elasticsearch==8.11.0

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2

# Deep Learning (for VGG)
tensorflow==2.15.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1

# CORS
python-jose[cryptography]==3.3.0
